+++ CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-01-20 18:32:33
+
+### Changed Files
+- `src/features/landing/components/HeroSection.tsx`
+
+### Statistics
+```
+ 1 file changed, 27 insertions(+)
+```
+
 ## [Checkpoint] - 2026-01-20 15:18:21
 
 ### Changed Files